requirements.txt
alchemy-logging==1.4.0
chardet==5.2.0
datasets==2.15.0
elasticsearch==8.13.0
Jinja2==3.1.6
kafka==1.3.5
pandarallel==1.6.5
pandas==2.2.2
patool==1.15.0
python-dotenv==1.0.1
python_dateutil==2.8.2
pyjnius==1.6.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
statsmodels==0.14.0
tqdm==4.66.1
transformers==4.56.2
openpyxl==3.1.4
numpy==1.26.4
setuptools==70.0.0
gliner2==1.1.2
drain3==0.9.11
click==8.3.1
torch==2.8.0